pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["UIika <1lys.rynd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = {extras = ["all"], version = "^0.115.8"}
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
google-genai = "^1.3.0"
markdown = "^3.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
